requirements.txt
discord.py==1.7.3
texttable==1.6.3
Hypercorn==0.11.2
numpy==1.21.0
Quart==0.15.1
confuse==1.5.0
discord==1.7.3
Pillow==8.3.0
tinydb==4.5.0
